requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0,<4.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
python-decouple>=3.8
psycopg2-binary>=2.9.9  # for PostgreSQL, remove if not using Postgres
gunicorn>=21.2.0        # for production (optional)
